+++ science-gateway/apps/executor/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "api"
+name = "executor"
 version = "0.1.0"
 edition = "2024"
