+++ science-gateway/apps/executor/src/core/entities/artifacts.rs
