+++ science-gateway/apps/executor/src/main.rs
@@ -1,0 +1,7 @@
+mod adapters;
+mod core;
+
+#[tokio::main]
+async fn main() {
+    adapters::rest::serve().await;
+}
