+++ science-gateway/apps/workbench/next-env.d.ts
