+++ science-gateway/apps/workbench/next.config.ts
